+++ src/main/scala/com/huemulsolutions/bigdata/test/Proc_PlanPruebas_Errores.scala
@@ -31,14 +31,19 @@ object Proc_PlanPruebas_Errores {
       val TablaMaster = new tbl_DatosBasicosErrores(huemulLib, Control)      
       TablaMaster.DataFramehuemul.setDataFrame(DF_RAW.DataFramehuemul.DataFrame, "DF_Original")
       
-      //BORRA HDFS ANTIGUO PARA EFECTOS DEL PLAN DE PRUEBAS
+   //BORRA HDFS ANTIGUO PARA EFECTOS DEL PLAN DE PRUEBAS
+      val a = huemulLib.spark.catalog.listTables(TablaMaster.GetCurrentDataBase()).collect()
+      if (a.filter { x => x.name.toUpperCase() == TablaMaster.TableName.toUpperCase()  }.length > 0) {
+        huemulLib.spark.sql(s"drop table if exists ${TablaMaster.GetTable()} ")
+      } 
+      
       val FullPath = new org.apache.hadoop.fs.Path(s"${TablaMaster.GetFullNameWithPath()}")
-      val fs = FileSystem.get(huemulLib.spark.sparkContext.hadoopConfiguration)       
+      val fs = FileSystem.get(huemulLib.spark.sparkContext.hadoopConfiguration) 
       if (fs.exists(FullPath))
         fs.delete(FullPath, true)
-      huemulLib.spark.sql(s"drop table if exists ${TablaMaster.GetTable()} ")
+        
    //BORRA HDFS ANTIGUO PARA EFECTOS DEL PLAN DE PRUEBAS
-      
+        
       TablaMaster.TipoValor.SetMapping("TipoValor")
       TablaMaster.Column_DQ_MaxDateTimeValue.SetMapping("timeStampValue")
       TablaMaster.Column_DQ_MaxDecimalValue.SetMapping("DecimalValue")
@@ -55,86 +60,113 @@ object Proc_PlanPruebas_Errores {
       //TODO: cambiar el parámetro "true" por algo.UPDATE O algo.NOUPDATE (en replaceValueOnUpdate
       Control.NewStep("Ejecución")
       TablaMaster.executeFull("DF_Final")
-      
+      var IdTestPlan: String = ""
       //Column_DQ_MinLen
       var ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_DQ_MinLen" && x.DQ_ErrorCode == 1020 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinLen", "No hay error encontrado", "ErrorCode = 1020", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinLen", "No hay error encontrado", "ErrorCode = 1020", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("DQ_MinLen", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinLen Codigo", "Errores encontrados correctamente", "ErrorCode = 1020, N° registros con error: 3", s"ErrorCode = 1020, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 3)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinLen Codigo", "Errores encontrados correctamente", "ErrorCode = 1020, N° registros con error: 3", s"ErrorCode = 1020, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 3)
+        Control.RegisterTestPlanFeature("DQ_MinLen", IdTestPlan)
       }
       
       //Column_DQ_MaxLen
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_DQ_MaxLen" && x.DQ_ErrorCode == 1020 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxLen", "No hay error encontrado", "ErrorCode = 1020", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxLen", "No hay error encontrado", "ErrorCode = 1020", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("DQ_MaxLen", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxLen Codigo", "Errores encontrados correctamente", "ErrorCode = 1020, N° registros con error: 2", s"ErrorCode = 1020, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxLen Codigo", "Errores encontrados correctamente", "ErrorCode = 1020, N° registros con error: 2", s"ErrorCode = 1020, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        Control.RegisterTestPlanFeature("DQ_MaxLen", IdTestPlan)
       }
       
       //Column_DQ_MinDecimalValue
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_DQ_MinDecimalValue" && x.DQ_ErrorCode == 1021 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDecimalValue", "No hay error encontrado", "ErrorCode = 1021", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDecimalValue", "No hay error encontrado", "ErrorCode = 1021", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("DQ_MinDecimalValue", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDecimalValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1021, N° registros con error: 2", s"ErrorCode = 1021, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDecimalValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1021, N° registros con error: 2", s"ErrorCode = 1021, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        Control.RegisterTestPlanFeature("DQ_MinDecimalValue", IdTestPlan)
       }
       
       //Column_DQ_MaxDecimalValue
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_DQ_MaxDecimalValue" && x.DQ_ErrorCode == 1021 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDecimalValue", "No hay error encontrado", "ErrorCode = 1021", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDecimalValue", "No hay error encontrado", "ErrorCode = 1021", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("DQ_MaxDecimalValue", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDecimalValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1021, N° registros con error: 1", s"ErrorCode = 1021, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 1)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDecimalValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1021, N° registros con error: 1", s"ErrorCode = 1021, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 1)
+        Control.RegisterTestPlanFeature("DQ_MaxDecimalValue", IdTestPlan)
       }
       
       //Column_DQ_MinDateTimeValue
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_DQ_MinDateTimeValue" && x.DQ_ErrorCode == 1022 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDateTimeValue", "No hay error encontrado", "ErrorCode = 1022", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDateTimeValue", "No hay error encontrado", "ErrorCode = 1022", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("DQ_MinDateTimeValue", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDateTimeValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1022, N° registros con error: 3", s"ErrorCode = 1022, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 3)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MinDateTimeValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1022, N° registros con error: 3", s"ErrorCode = 1022, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 3)
+        Control.RegisterTestPlanFeature("DQ_MinDateTimeValue", IdTestPlan)
       }
       
       //Column_DQ_MaxDateTimeValue
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_DQ_MaxDateTimeValue" && x.DQ_ErrorCode == 1022 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDateTimeValue", "No hay error encontrado", "ErrorCode = 1022", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDateTimeValue", "No hay error encontrado", "ErrorCode = 1022", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("DQ_MaxDateTimeValue", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDateTimeValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1022, N° registros con error: 2", s"ErrorCode = 1022, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_DQ_MaxDateTimeValue Codigo", "Errores encontrados correctamente", "ErrorCode = 1022, N° registros con error: 2", s"ErrorCode = 1022, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        Control.RegisterTestPlanFeature("DQ_MaxDateTimeValue", IdTestPlan)
       }
       
       //Column_NotNull
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_NotNull" && x.DQ_ErrorCode == 1023 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_NotNull", "No hay error encontrado", "ErrorCode = 1023", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0) {
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_NotNull", "No hay error encontrado", "ErrorCode = 1023", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("Nullable", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_NotNull Codigo", "Errores encontrados correctamente", "ErrorCode = 1023, N° registros con error: 1", s"ErrorCode = 1023, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 1)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_NotNull Codigo", "Errores encontrados correctamente", "ErrorCode = 1023, N° registros con error: 1", s"ErrorCode = 1023, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 1)
+        Control.RegisterTestPlanFeature("Nullable", IdTestPlan)
       }
       
       //Column_IsUnique
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.ColumnName == "Column_IsUnique" && x.DQ_ErrorCode == 2006 }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_IsUnique", "No hay error encontrado", "ErrorCode = 2006", s"SIN REGISTRO", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_IsUnique", "No hay error encontrado", "ErrorCode = 2006", s"SIN REGISTRO", false)
+        Control.RegisterTestPlanFeature("IsUnique", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg(0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_IsUnique Codigo", "Errores encontrados correctamente", "ErrorCode = 2006, N° registros Not Unique: 2", s"ErrorCode = 2006, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR Column_IsUnique Codigo", "Errores encontrados correctamente", "ErrorCode = 2006, N° registros Not Unique: 2", s"ErrorCode = 2006, N° registros con error: ${ErrorEncontrado.DQ_NumRowsError}", ErrorEncontrado.DQ_NumRowsError == 2)
+        Control.RegisterTestPlanFeature("IsUnique", IdTestPlan)
       }
       
       //N° Total de errores por DQ debe ser = 8
       ErrorReg = TablaMaster.DataFramehuemul.getDQResult().filter { x => x.DQ_IsError }
-      if (ErrorReg == null || ErrorReg.length == 0)
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR N° DQ fallados", "No hay error encontrado", "Cantidad de errores DQ = 8", s"Cantidad de errores DQ = ${ErrorReg.length}", false)
+      if (ErrorReg == null || ErrorReg.length == 0){
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR N° DQ fallados", "No hay error encontrado", "Cantidad de errores DQ = 8", s"Cantidad de errores DQ = ${ErrorReg.length}", false)
+        Control.RegisterTestPlanFeature("executeFull", IdTestPlan)
+      }
       else {
         val ErrorEncontrado = ErrorReg.length
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR N° DQ Fallados cod", "Errores encontrados correctamente", "Cantidad de errores DQ = 8", s"Cantidad de errores DQ = ${ErrorEncontrado}", ErrorEncontrado == 8)
+        IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR N° DQ Fallados cod", "Errores encontrados correctamente", "Cantidad de errores DQ = 8", s"Cantidad de errores DQ = ${ErrorEncontrado}", ErrorEncontrado == 8)
+        Control.RegisterTestPlanFeature("executeFull", IdTestPlan)
       }
       
       println(s"detalle del 100% de DQ aplicado en esta prueba")
@@ -146,7 +178,8 @@ object Proc_PlanPruebas_Errores {
           Control.FinishProcessOK
     } catch {
       case e: Exception => 
-        Control.RegisterTestPlan(TestPlanGroup, "ERROR", "ERROR DE PROGRAMA -  no deberia tener errror", "sin error", s"con error", false)
+        val IdTestPlan = Control.RegisterTestPlan(TestPlanGroup, "ERROR", "ERROR DE PROGRAMA -  no deberia tener errror", "sin error", s"con error", false)
+        Control.RegisterTestPlanFeature("executeFull", IdTestPlan)
         Control.Control_Error.GetError(e, this.getClass.getSimpleName, null)
         Control.FinishProcessError()
     }
